@@ -1,12 +1,16 @@
 module main
 
-go 1.20
+go 1.21
 
 require (
-	github.com/aws/aws-sdk-go v1.43.9
-	github.com/salrashid123/aws_hmac/vault v0.0.0
-	github.com/salrashid123/aws_hmac/vault/v4 v4.0.0
+	github.com/aws/aws-sdk-go v1.45.3
+	github.com/salrashid123/aws_hmac/stsschema v0.0.0
+	github.com/salrashid123/aws_hmac/vault v0.0.0-00010101000000-000000000000
+	github.com/salrashid123/aws_hmac/vault/signer v0.0.0
+	github.com/salrashid123/aws_hmac/vault/signer/v4 v4.0.0
 )
+
+require github.com/gorilla/schema v1.2.0
 
 require (
 	github.com/aws/smithy-go v1.14.2 // indirect
@@ -21,7 +25,7 @@ require (
 	github.com/hashicorp/go-secure-stdlib/strutil v0.1.2 // indirect
 	github.com/hashicorp/go-sockaddr v1.0.2 // indirect
 	github.com/hashicorp/hcl v1.0.0 // indirect
-	github.com/hashicorp/vault/api v1.9.2 // indirect
+	github.com/hashicorp/vault/api v1.10.0 // indirect
 	github.com/jmespath/go-jmespath v0.4.0 // indirect
 	github.com/mitchellh/go-homedir v1.1.0 // indirect
 	github.com/mitchellh/mapstructure v1.5.0 // indirect
@@ -33,6 +37,8 @@ require (
 )
 
 replace (
+	github.com/salrashid123/aws_hmac/stsschema => ../../stsschema
 	github.com/salrashid123/aws_hmac/vault => ../../vault
-	github.com/salrashid123/aws_hmac/vault/v4 => ../../vault/v4
+	github.com/salrashid123/aws_hmac/vault/signer => ../../vault/signer
+	github.com/salrashid123/aws_hmac/vault/signer/v4 => ../../vault/signer/v4
 )
