@@ -1,10 +1,18 @@
 module github.com/salrashid123/aws_hmac/vault
 
-go 1.20
-
-require github.com/hashicorp/vault/api v1.9.2
+go 1.21
 
 require (
+	github.com/aws/aws-sdk-go v1.45.3
+	github.com/gorilla/schema v1.2.0
+	github.com/salrashid123/aws_hmac/stsschema v0.0.0-00010101000000-000000000000
+	github.com/salrashid123/aws_hmac/vault/signer v0.0.0
+	github.com/salrashid123/aws_hmac/vault/signer/v4 v4.0.0
+
+)
+
+require (
+	github.com/aws/smithy-go v1.14.2 // indirect
 	github.com/cenkalti/backoff/v3 v3.0.0 // indirect
 	github.com/go-jose/go-jose/v3 v3.0.0 // indirect
 	github.com/hashicorp/errwrap v1.1.0 // indirect
@@ -16,6 +24,7 @@ require (
 	github.com/hashicorp/go-secure-stdlib/strutil v0.1.2 // indirect
 	github.com/hashicorp/go-sockaddr v1.0.2 // indirect
 	github.com/hashicorp/hcl v1.0.0 // indirect
+	github.com/hashicorp/vault/api v1.10.0 // indirect
 	github.com/mitchellh/go-homedir v1.1.0 // indirect
 	github.com/mitchellh/mapstructure v1.5.0 // indirect
 	github.com/ryanuber/go-glob v1.0.0 // indirect
@@ -23,4 +32,10 @@ require (
 	golang.org/x/net v0.7.0 // indirect
 	golang.org/x/text v0.7.0 // indirect
 	golang.org/x/time v0.0.0-20200416051211-89c76fbcd5d1 // indirect
+)
+
+replace (
+	github.com/salrashid123/aws_hmac/stsschema => ../stsschema
+	github.com/salrashid123/aws_hmac/vault/signer => ./signer
+	github.com/salrashid123/aws_hmac/vault/signer/v4 => ./signer/v4
 )
