@@ -3,23 +3,36 @@ package main
 import (
 	"bytes"
 	"context"
+	"crypto/sha256"
 	"encoding/binary"
+	"encoding/hex"
+	"encoding/xml"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"log"
 	"net/http"
+	"net/url"
 	"strings"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	v4 "github.com/aws/aws-sdk-go/aws/signer/v4"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/sts"
 
+	"github.com/aws/aws-sdk-go/aws/credentials"
+	awssession "github.com/aws/aws-sdk-go/aws/session"
+	v4 "github.com/aws/aws-sdk-go/aws/signer/v4"
 	"github.com/miekg/pkcs11"
 
 	"flag"
 
+	"github.com/gorilla/schema"
+
+	"github.com/salrashid123/aws_hmac/stsschema"
+
 	hmaccred "github.com/salrashid123/aws_hmac/pkcs"
-	hmacsigner "github.com/salrashid123/aws_hmac/pkcs/v4"
+	hmacsigner "github.com/salrashid123/aws_hmac/pkcs/signer"
+	hmacsignerv4 "github.com/salrashid123/aws_hmac/pkcs/signer/v4"
 )
 
 const (
@@ -29,6 +42,8 @@ const (
 	awsDateHeader          = "x-amz-date"
 	awsTimeFormatLong      = "20060102T150405Z"
 	awsTimeFormatShort     = "20060102"
+
+	emptyPayloadHash = "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
 )
 
 var (
@@ -37,7 +52,9 @@ var (
 	accessKeyID     = flag.String("accessKeyID", "", "AWS AccessKeyID")
 	secretAccessKey = flag.String("secretAccessKey", "", "AWS SecretAccessKey")
 
-	awsRegion = flag.String("awsRegion", "us-east-1", "AWS Region")
+	awsRegion       = flag.String("awsRegion", "us-east-1", "AWS Region")
+	roleARN         = flag.String("roleARN", "arn:aws:iam::291738886548:role/gcpsts", "Role to assume")
+	roleSessionName = "mysession"
 )
 
 func main() {
@@ -49,11 +66,11 @@ func main() {
 
 	// // // **************  STS
 
-	log.Println("Using  Standard AWS v4Signer")
+	log.Println("Using Default AWS v4Signer and StaticCredentials to make REST GET call to GetCallerIdentity")
 	creds := credentials.NewStaticCredentials(*accessKeyID, *secretAccessKey, "")
-	rbody := strings.NewReader("")
 	signer := v4.NewSigner(creds)
-	req, err := http.NewRequest(http.MethodPost, "https://sts.amazonaws.com?Action=GetCallerIdentity&Version=2011-06-15", rbody)
+	rbody := strings.NewReader("")
+	req, err := http.NewRequest(http.MethodGet, "https://sts.amazonaws.com?Action=GetCallerIdentity&Version=2011-06-15", rbody)
 	if err != nil {
 		log.Fatalln(err)
 	}
@@ -79,6 +96,8 @@ func main() {
 	log.Printf("   Response using AWS STS NewStaticCredentials and Standard v4.Singer \n%s", string(b))
 
 	// ***************************************************************************
+
+	log.Println("   Initializing PKCS Keyset embedding AWS Secret")
 
 	pin := "mynewpin"
 	// Init PKCS
@@ -158,8 +177,10 @@ func main() {
 	p.Finalize()
 	p.Destroy()
 
-	cc, err := hmaccred.NewHMACCredential(&hmaccred.HMACCredentialConfig{
-		PKCSConfig: hmaccred.PKCSConfig{
+	ctx := context.Background()
+
+	pkcsSigner, err := hmacsigner.NewPKCSSigner(&hmacsigner.PKCSSignerConfig{
+		PKCSConfig: hmacsigner.PKCSConfig{
 			Library: "/usr/lib/x86_64-linux-gnu/softhsm/libsofthsm2.so",
 			Slot:    0,
 			Label:   "HMACKey",
@@ -169,46 +190,120 @@ func main() {
 		AccessKeyID: *accessKeyID,
 	})
 	if err != nil {
-		log.Fatalln(err)
+		log.Fatalf("%v", err)
 	}
+	hmacSigner := hmacsignerv4.NewSigner()
 
-	// now call the endpoint
-
-	ctx := context.Background()
-	hs := hmacsigner.NewSigner()
-
-	body := strings.NewReader("")
-	sreq, err := http.NewRequest(http.MethodPost, "https://sts.amazonaws.com?Action=GetCallerIdentity&Version=2011-06-15", body)
+	sessionCredentials, err := hmaccred.NewAWSPKCSCredentials(hmaccred.PKCSProvider{
+		GetSessionTokenInput: &stsschema.GetSessionTokenInput{
+			DurationSeconds: aws.Int64(3600),
+		},
+		Version:    "2011-06-15",
+		Region:     *awsRegion,
+		PKCSSigner: pkcsSigner,
+	})
 	if err != nil {
-		log.Fatalln(err)
+		log.Fatalf("Could not initialize Tink Credentials %v", err)
 	}
 
-	// $ touch empty.txt
-	// $ sha256sum empty.txt
-	//   e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855  empty.txt
+	assumeRoleCredentials, err := hmaccred.NewAWSPKCSCredentials(hmaccred.PKCSProvider{
+		AssumeRoleInput: &stsschema.AssumeRoleInput{
+			RoleArn:         aws.String(*roleARN),
+			RoleSessionName: aws.String(roleSessionName),
+			DurationSeconds: aws.Int64(3600),
+		},
+		Version:    "2011-06-15",
+		Region:     *awsRegion,
+		PKCSSigner: pkcsSigner,
+	})
+	if err != nil {
+		log.Fatalf("Could not read response Body%v", err)
+	}
+	// *******************************************
 
-	payloadHash := "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
-	hs.SignHTTP(ctx, *cc, sreq, payloadHash, "sts", *awsRegion, time.Now())
+	fmt.Println("-------------------------------- Calling HTTP POST on  GetCallerIdentity using Tink Signer")
+
+	getCallerIdentityRequestStruct := stsschema.GetCallerIdentityRequest{
+		Action:  "GetCallerIdentity",
+		Version: "2011-06-15",
+	}
+
+	postForm := url.Values{}
+	err = schema.NewEncoder().Encode(getCallerIdentityRequestStruct, postForm)
 	if err != nil {
 		log.Fatalf("%v", err)
 	}
 
-	log.Printf("    Signed RequestURI: %v\n", sreq.RequestURI)
+	getCallerIdentityPostRequest, err := http.NewRequest(http.MethodPost, "https://sts.amazonaws.com", strings.NewReader(postForm.Encode()))
+	if err != nil {
+		log.Fatalf("%v", err)
+	}
+	getCallerIdentityPostRequest.Header.Add("Content-Type", "application/x-www-form-urlencoded")
 
-	sres, err := http.DefaultClient.Do(sreq)
+	hasher := sha256.New()
+	hasher.Write([]byte(postForm.Encode()))
+	postPayloadHash := hex.EncodeToString(hasher.Sum(nil))
+
+	hmacSigner.SignHTTP(ctx, *pkcsSigner, getCallerIdentityPostRequest, postPayloadHash, "sts", *awsRegion, time.Now())
 	if err != nil {
 		log.Fatalf("%v", err)
 	}
 
-	defer sres.Body.Close()
-	if sres.StatusCode != 200 {
-		log.Printf("Response is not 200 %v\n", err)
+	getCallerIdentityPostResponse, err := http.DefaultClient.Do(getCallerIdentityPostRequest)
+	if err != nil {
+		log.Fatalf("%v", err)
 	}
-	sb, err := ioutil.ReadAll(sres.Body)
+	defer getCallerIdentityPostResponse.Body.Close()
+
+	if getCallerIdentityPostResponse.StatusCode != 200 {
+		bodyBytes, err := io.ReadAll(getCallerIdentityPostResponse.Body)
+		if err != nil {
+			log.Fatal(err)
+		}
+		log.Fatalf("Response is not 200 %s %v\n", string(bodyBytes), err)
+	}
+
+	getResponseData, err := ioutil.ReadAll(getCallerIdentityPostResponse.Body)
+	if err != nil {
+		log.Fatalf("Could not read response Body%v", err)
+	}
+	var getCallerIdentityResponseStruct stsschema.GetCallerIdentityResponse
+	err = xml.Unmarshal(getResponseData, &getCallerIdentityResponseStruct)
+	if err != nil {
+		log.Fatalf("Could not read response Body%v", err)
+	}
+	log.Printf("GetCallerIdentityResponse UserID %s\n", getCallerIdentityResponseStruct.CallerIdentityResult.UserId)
+
+	fmt.Println("-------------------------------- GetCallerIdentity with SessionToken SDK")
+
+	sess, err := awssession.NewSession(&aws.Config{
+		Region:      aws.String(*awsRegion),
+		Credentials: sessionCredentials,
+	})
+
+	stssvc := sts.New(sess, aws.NewConfig().WithRegion(*awsRegion))
+	stsresp, err := stssvc.GetCallerIdentity(&sts.GetCallerIdentityInput{})
+	if err != nil {
+		log.Fatalf("Could not read response Body%v", err)
+	}
+	log.Printf("STS Identity from API %s\n", *stsresp.UserId)
+
+	fmt.Println("-------------------------------- GetCallerIdentity with AssumeRole SDK")
+
+	sess2, err := awssession.NewSession(&aws.Config{
+		Region:      aws.String(*awsRegion),
+		Credentials: assumeRoleCredentials,
+	})
+
 	if err != nil {
 		log.Fatalf("Could not read response Body%v", err)
 	}
 
-	log.Printf("    STS Response:  \n%s", string(sb))
+	stssvc2 := sts.New(sess2, aws.NewConfig().WithRegion(*awsRegion))
+	stsresp2, err := stssvc2.GetCallerIdentity(&sts.GetCallerIdentityInput{})
+	if err != nil {
+		log.Fatalf("Could not read response Body%v", err)
+	}
+	log.Printf("Assumed role ARN: %s\n", *stsresp2.Arn)
 
 }
