@@ -3,19 +3,31 @@ package main
 import (
 	"bytes"
 	"context"
+	"crypto/sha256"
+	"encoding/hex"
 	"encoding/json"
+	"encoding/xml"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"log"
 	"math/rand"
 	"net/http"
+	"net/url"
 	"strings"
 	"time"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/sts"
+	"github.com/gorilla/schema"
 
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	v4 "github.com/aws/aws-sdk-go/aws/signer/v4"
 
 	"flag"
+
+	"github.com/salrashid123/aws_hmac/stsschema"
 
 	"github.com/golang/protobuf/proto"
 	"github.com/google/tink/go/core/registry"
@@ -27,7 +39,8 @@ import (
 	tinkpb "github.com/google/tink/go/proto/tink_go_proto"
 
 	hmaccred "github.com/salrashid123/aws_hmac/tink"
-	hmacsigner "github.com/salrashid123/aws_hmac/tink/v4"
+	hmacsigner "github.com/salrashid123/aws_hmac/tink/signer"
+	hmacsignerv4 "github.com/salrashid123/aws_hmac/tink/signer/v4"
 )
 
 const (
@@ -45,31 +58,33 @@ const (
 	RawPrefix        = ""
 
 	tagSize = 32
+
+	// $ touch empty.txt
+	// $ sha256sum empty.txt
+	//   e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855  empty.txt
+
+	emptyPayloadHash = "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
 )
 
 var (
 	accessKeyID     = flag.String("accessKeyID", "", "AWS AccessKeyID")
 	secretAccessKey = flag.String("secretAccessKey", "", "AWS SecretAccessKey")
-
-	awsRegion = flag.String("awsRegion", "us-east-1", "AWS Region")
-
-	keyURI = flag.String("keyURI", "projects/mineral-minutia-820/locations/us-central1/keyRings/mykeyring/cryptoKeys/key1", "KMS Key Uri")
+	awsRegion       = flag.String("awsRegion", "us-east-1", "AWS Region")
+	keyURI          = flag.String("keyURI", "projects/PROJECT_ID/locations/us-central1/keyRings/mykeyring/cryptoKeys/key1", "KMS Key Uri")
+	roleARN         = flag.String("roleARN", "arn:aws:iam::291738886548:role/gcpsts", "Role to assume")
+	roleSessionName = "mysession"
 )
 
 func main() {
 	flag.Parse()
 
-	if *accessKeyID == "" || *secretAccessKey == "" {
-		log.Fatal("accessKeyID and secretAccessKey must be set")
-	}
-
 	// // // **************  STS
 
-	log.Println("Using  Standard AWS v4Signer")
+	log.Println("Using Default AWS v4Signer and StaticCredentials to make REST GET call to GetCallerIdentity")
 	creds := credentials.NewStaticCredentials(*accessKeyID, *secretAccessKey, "")
-	rbody := strings.NewReader("")
 	signer := v4.NewSigner(creds)
-	req, err := http.NewRequest(http.MethodPost, "https://sts.amazonaws.com?Action=GetCallerIdentity&Version=2011-06-15", rbody)
+	rbody := strings.NewReader("")
+	req, err := http.NewRequest(http.MethodGet, "https://sts.amazonaws.com?Action=GetCallerIdentity&Version=2011-06-15", rbody)
 	if err != nil {
 		log.Fatalln(err)
 	}
@@ -96,11 +111,12 @@ func main() {
 
 	// ***************************************************************************
 
+	log.Println("   Initializing GCP KMS Encrypted Tink Keyset embedding AWS Secret")
+
 	if *keyURI == "" {
 		log.Fatal("keyURI must be set if mode=tink")
 	}
 
-	log.Println("   Create tink subtle.HMAC using secret")
 	id := rand.Uint32()
 
 	tk, err := subtle.NewHMAC(common_go_proto.HashType_name[int32(common_go_proto.HashType_SHA256)], []byte("AWS4"+*secretAccessKey), tagSize)
@@ -116,7 +132,7 @@ func main() {
 		},
 		KeyValue: tk.Key,
 	}
-	// log.Printf("    Tink HmacKey Key: %v", base64.StdEncoding.EncodeToString(k.GetKeyValue()))
+	//log.Printf("    Tink HmacKey Key: %v", base64.StdEncoding.EncodeToString(k.GetKeyValue()))
 
 	keyserialized, err := proto.Marshal(k)
 	if err != nil {
@@ -210,52 +226,133 @@ func main() {
 	}
 	log.Println("   Tink Keyset:\n", string(prettyJSON.Bytes()))
 
-	cc, err := hmaccred.NewHMACCredential(&hmaccred.HMACCredentialConfig{
-		TinkConfig: hmaccred.TinkConfig{
+	// ********************************************
+
+	// ********************************************
+
+	ctx := context.Background()
+
+	tinkSigner, err := hmacsigner.NewTinkSigner(&hmacsigner.TinkSignerConfig{
+		TinkConfig: hmacsigner.TinkConfig{
 			KmsBackend: backend,
 			JSONBytes:  prettyJSON.Bytes(),
 		},
 		AccessKeyID: *accessKeyID,
 	})
 	if err != nil {
-		log.Fatalln(err)
+		log.Fatalf("%v", err)
 	}
+	hmacSigner := hmacsignerv4.NewSigner()
 
-	ctx := context.Background()
-	hs := hmacsigner.NewSigner()
-
-	body := strings.NewReader("")
-	sreq, err := http.NewRequest(http.MethodPost, "https://sts.amazonaws.com?Action=GetCallerIdentity&Version=2011-06-15", body)
+	sessionCredentials, err := hmaccred.NewAWSTinkCredentials(hmaccred.TINKProvider{
+		GetSessionTokenInput: &stsschema.GetSessionTokenInput{
+			DurationSeconds: aws.Int64(3600),
+		},
+		Version:    "2011-06-15",
+		Region:     *awsRegion,
+		TinkSigner: tinkSigner,
+	})
 	if err != nil {
-		log.Fatalln(err)
+		log.Fatalf("Could not initialize Tink Credentials %v", err)
 	}
 
-	// $ touch empty.txt
-	// $ sha256sum empty.txt
-	//   e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855  empty.txt
+	assumeRoleCredentials, err := hmaccred.NewAWSTinkCredentials(hmaccred.TINKProvider{
+		AssumeRoleInput: &stsschema.AssumeRoleInput{
+			RoleArn:         aws.String(*roleARN),
+			RoleSessionName: aws.String(roleSessionName),
+			DurationSeconds: aws.Int64(3600),
+		},
+		Version:    "2011-06-15",
+		Region:     *awsRegion,
+		TinkSigner: tinkSigner,
+	})
+	if err != nil {
+		log.Fatalf("Could not read response Body%v", err)
+	}
+	// *******************************************
 
-	payloadHash := "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
-	hs.SignHTTP(ctx, *cc, sreq, payloadHash, "sts", *awsRegion, time.Now())
+	fmt.Println("-------------------------------- Calling HTTP POST on  GetCallerIdentity using Tink Signer")
+
+	getCallerIdentityRequestStruct := stsschema.GetCallerIdentityRequest{
+		Action:  "GetCallerIdentity",
+		Version: "2011-06-15",
+	}
+
+	postForm := url.Values{}
+	err = schema.NewEncoder().Encode(getCallerIdentityRequestStruct, postForm)
 	if err != nil {
 		log.Fatalf("%v", err)
 	}
 
-	log.Printf("    Signed RequestURI: %v\n", sreq.RequestURI)
+	getCallerIdentityPostRequest, err := http.NewRequest(http.MethodPost, "https://sts.amazonaws.com", strings.NewReader(postForm.Encode()))
+	if err != nil {
+		log.Fatalf("%v", err)
+	}
+	getCallerIdentityPostRequest.Header.Add("Content-Type", "application/x-www-form-urlencoded")
 
-	sres, err := http.DefaultClient.Do(sreq)
+	hasher := sha256.New()
+	hasher.Write([]byte(postForm.Encode()))
+	postPayloadHash := hex.EncodeToString(hasher.Sum(nil))
+
+	hmacSigner.SignHTTP(ctx, *tinkSigner, getCallerIdentityPostRequest, postPayloadHash, "sts", *awsRegion, time.Now())
 	if err != nil {
 		log.Fatalf("%v", err)
 	}
 
-	defer sres.Body.Close()
-	if sres.StatusCode != 200 {
-		log.Printf("Response is not 200 %v\n", err)
+	getCallerIdentityPostResponse, err := http.DefaultClient.Do(getCallerIdentityPostRequest)
+	if err != nil {
+		log.Fatalf("%v", err)
 	}
-	sb, err := ioutil.ReadAll(sres.Body)
+	defer getCallerIdentityPostResponse.Body.Close()
+
+	if getCallerIdentityPostResponse.StatusCode != 200 {
+		bodyBytes, err := io.ReadAll(getCallerIdentityPostResponse.Body)
+		if err != nil {
+			log.Fatal(err)
+		}
+		log.Fatalf("Response is not 200 %s %v\n", string(bodyBytes), err)
+	}
+
+	getResponseData, err := ioutil.ReadAll(getCallerIdentityPostResponse.Body)
+	if err != nil {
+		log.Fatalf("Could not read response Body%v", err)
+	}
+	var getCallerIdentityResponseStruct stsschema.GetCallerIdentityResponse
+	err = xml.Unmarshal(getResponseData, &getCallerIdentityResponseStruct)
+	if err != nil {
+		log.Fatalf("Could not read response Body%v", err)
+	}
+	log.Printf("GetCallerIdentityResponse UserID %s\n", getCallerIdentityResponseStruct.CallerIdentityResult.UserId)
+
+	fmt.Println("-------------------------------- GetCallerIdentity with SessionToken SDK")
+
+	sess, err := session.NewSession(&aws.Config{
+		Region:      aws.String(*awsRegion),
+		Credentials: sessionCredentials,
+	})
+
+	stssvc := sts.New(sess, aws.NewConfig().WithRegion(*awsRegion))
+	stsresp, err := stssvc.GetCallerIdentity(&sts.GetCallerIdentityInput{})
+	if err != nil {
+		log.Fatalf("Could not read response Body%v", err)
+	}
+	log.Printf("STS Identity from API %s\n", *stsresp.UserId)
+
+	fmt.Println("-------------------------------- GetCallerIdentity with AssumeRole SDK")
+
+	sess2, err := session.NewSession(&aws.Config{
+		Region:      aws.String(*awsRegion),
+		Credentials: assumeRoleCredentials,
+	})
+
 	if err != nil {
 		log.Fatalf("Could not read response Body%v", err)
 	}
 
-	log.Printf("    STS Response:  \n%s", string(sb))
-
+	stssvc2 := sts.New(sess2, aws.NewConfig().WithRegion(*awsRegion))
+	stsresp2, err := stssvc2.GetCallerIdentity(&sts.GetCallerIdentityInput{})
+	if err != nil {
+		log.Fatalf("Could not read response Body%v", err)
+	}
+	log.Printf("Assumed role ARN: %s\n", *stsresp2.Arn)
 }
